Renamed UGen._new() to UGen._new_expanded().
UGen._new_expanded() now delegates instantiating UGens to UGen._new_single().
This in turn allows us to override the instantiation procedure for a single
UGen (post-input-expansion) in order to decide whether to return an instance of
that class, or an instance of a different class.

See Sum3 and Sum4 for an example of this in action: both classes override their
_new_single() class method in order to provide optimized alternate
instantiations for "degenerate" input.